requirements.txt
# Core
flask>=2.2,<3.0
python-dotenv>=1.0.0
pypdf>=3.9.0

# Embeddings
openai>=1.0.0

# LangChain ecosystem (aligned on 0.3.x)
langchain==0.3.7
langchain-community==0.3.6
langchain-openai>=0.2.0
langchain-pinecone==0.2.12
langchain-experimental>=0.0.80
langchain-huggingface==0.1.0
langchain-text-splitters>=0.3.0,<0.4.0

# Pinecone client
pinecone>=6.0.0,<8.0.0

# Mongo
pymongo>=4.0.0

# Deployment
gunicorn>=21.2.0